fix: Parented in-scene objects under dynamically spawned objects are destroyed when the network session ends but original scene remains loaded. [MTT-7560] (#2737)
* fix

Fixes the issue with an in-scene placed NetworkObject being parented under a dynamically spawned NetworkObject and a sudden shutdown/exit occurs. This will remove the child in-scene placed NetworkObject from the parent during the NetworkManager shutdown process where it destroys all dynamically spawned NetworkObjects.'

Adjusting the parenting rules slightly to only deny parenting if the child or parent is not spawned.


* test

updating the test to reflect the updated error message when a client attempts to destroy a spawned NetworkObject while still connected to a session.
Updating the scene objects not destroyed test.

Adding `ChildSceneObjectsDoNotDestroyOnShutdown`

Adding ChildSceneObjectsDoNotDestroyOnShutdown test to validate this PR.
Updating ParentedNetworkTransformTest to provide enough time to let interpolation to finish.