fix & test
Fixing some additional issues with Rigidbody testing and assuring we are comparing the correct position and rotation. With NetworkTransformOwnershipTests, it turned out that the movement of the non-kinematic body was giving the Rigidbody a linear velocity under certain conditions.

Removing the warning about resetting dirty when not initialized during NetworkVariableTests (removing a couple thousand log messages).

Since the service is now consuming the initial synchronize event, updated how in-scene objects are spawned initially.