test
Realized we needed an overhaul with our rigid body tests so I did a bit of refactoring.
Added temporary internal method, OnSetupRigidbody, to both network rigid body components and will circle back to this test once the prespawn PR is merged to replace it with the NetworkBehaviour.OnNetworkPreSpawn.

Deleted the remaining of the completely commented out rigid body tests that are now being covered in the NetworkRigidbodyTests.