update
Decided that we don't need an extra class to handle this and to migrate the NetworkMessageTypes and type to NetworkMessageTypes enum index value table creation within the GetMessages method.

Removed the message ordering related tests since this it was no longer needed.

Adjusted name of integration test ILPPMessageProvider.IntegrationTestNoMessages to better reflect what it is used for and also removed it from the NetworkMessageManager.