update
Migrated non-authority NetworkTransform updates to the pre late update to more align with how it is handled in v1.x.
Added NetworkRigidbodyBase reference to NetworkObject.
For now, added the ability to add a fixed joint when parenting two NetworkObjects with NetworkTransforms and NetworkRigidbodies that both have UseRigidBodyForMotion set to true (i.e. Rigidbody has no concept of "local space" motion).