update
Unnamed and named messages now allow you to send a client id list of 1 on a client side if the client id is the server.

When registering a named message that is already registered, if the log mode is set to developer then a warning will be logged to the console informing the user they are overwriting the existing registration callback.