fix: Added support for byte on NetworkVariable through codegen (#2953)
* fix: Added support for byte on NetworkVariable through codegen

the codegen wasn't picking byte  - it's because a variable of a type other than NetworkVariable<byte> (such as NetworkVariable<FixedString32Bytes>) is serializing a byte value as part of its delta serialization. Delta serialization was added in NGO 1.9, which is why it doesn't happen in previous versions. This code assumes the existence of a byte serializer, but if the user doesn't have a NetworkVariable<byte> in their code, the byte serializer won't be initialized by the codegen pass.
credits: Kitty Drapper

* fix: Converted serialization operation to their appropriate Safe counterparts

Fixed safety issue that would trigger exception on runtime

* fix

Using WriteByteSafe as opposed to WriteValueSafe.

* update

adding changelog entry

---------

Co-authored-by: NoelStephensUnity <noel.stephens@unity3d.com>