fix: abide by the original RigidBody2D body type
fix: the 1.10 imposed Dynamic body type even when the user set it originally as Kinematic

Implemented a check that changes the body type to Dynamic when there is ownership only if originally it was instantiated as Dynamic, otherwise, it keeps it Kinematic. This to avoid overriding what the user had originally designed.