update
Renaming some internal properties that are common names and would conflict with serialization if a user defined them in a derived class.

Making sure both the position and rotation interpolators are configured for the transform space when reset.

Adding the ability to change the parent locally when running with a client-server network topology.
Adding SwitchTransformSpaceWhenParented property to the NetworkTransformEditor.
Shifting all of the non-authority's final synchronization logic to occur either once a client's initial connection synchronization is complete or when all of the NetworkObject's components have run through the spawn process.