fix: NetworkTransform initial synchronization, parenting, smooth transitions between transform spaces, and UI updates (#3013)
* fix

This fixes a synchronization issue with owner authoritative instances when running in client-server mode and parenting. When the server sends a parent sync message back with the position of the client's owned object it is out of sync with the client's position by the time the message is received. This fix gives users the option to not synchronize the transform position of a NetworkObject when using an owner authoritative motion mode.

* fix

This fix provides users with the option to have child `NetworkTransform`(s) tick synchronized with their parent. When the parent has `TickSyncChildren` enabled, any child `NetworkTransform`(s) that the local client has authority over will force a synchronization of their transform state. When a client is not the authority of the parent but it is the authority of one or more children, then when the parent sends a state update it will force the child/children the client has authority over to send a state update. Children will still send state updates if the transform has a delta that exceeds an axis threshold and the parent has yet to send a synchronization, but if a parent sends a synchronization within the same tick period the child/children will still send another state update.

* wip

Better approach, but still flawed.
I think converting any remaining items in the interpolators when the transform space changes will yield consistent results.

* fix

Seamless transform space transitions when parenting.

* update

We no longer need to send lossy scale since we defer applying the initial synchronization data until the non-authority side has finished synchronizing.

* fix

Fixing some minor issues with nested `NetworkTransforms` (i.e. under the same `NetworkObject`) and adding additional comments.
Reverting back to just checking `InLocalSpace` (as opposed to the interpolator's) when applying a position on the non-authority side.

* test fix

Fixing some issues with `NestedNetworkTransformTests` assets to account for some past updates.
Fixing an issue with `NetworkTransformTests` where it needs to not just advance the tick (for updates) but needs to also advance to the next tick.

* fix

caught scenario where `NetworkObject.NetworkTransforms` could not be set properly.

* fix

This fixes a minor "blip" when transitioning from world to local space,

* test

disabling two tests that need to be reviewed by Kitty and/or re-written.

* fix

Updating all NetworkTransform instances that belong to a specific NetworkObject as opposed to just the root.
Also not updating when not spawned or disabled NetworkTransforms.

* test

Making test asset adjustments for recent updates in NetworkTransform.

* update

Adding change log entry.

* update

mid-point check-in

* test

Updating interpolation test to recent NetworkTransform updates.

* update

wip towards getting a cleaner divide between the two modes.

* update

Some adjustments to when nested children check their own state updates or exit early.

* update

Cleaning up some Rigidbody references and making Rigidbody & Rigidbody2D accessible via NetworkRigidbody and NetworkRigidbody2D.

* update

Adding an internal InternalOnNetworkObjectParentChanged method that gets invoked ahead of OnNetworkObjectParentChanged so user code can't override any NGO components that need this kind of notification.

* update

Adding InternalOnNetworkSessionSynchronized to be invoked by NGO components to assure they don't get overridden by user script.

* update

Renaming some internal properties that are common names and would conflict with serialization if a user defined them in a derived class.

Making sure both the position and rotation interpolators are configured for the transform space when reset.

Adding the ability to change the parent locally when running with a client-server network topology.
Adding SwitchTransformSpaceWhenParented property to the NetworkTransformEditor.
Shifting all of the non-authority's final synchronization logic to occur either once a client's initial connection synchronization is complete or when all of the NetworkObject's components have run through the spawn process.

* update

Adding some improved and needed UI updates to NetworkBehaviour derived components and NetworkManager.
This includes a new NetcodeEditorBase class that allows users to more easily create customized, multi-generation, components that have their properties organized based on the generation of each child.

* fix

Wrapping the NetworkRigidbodyBaseEditor within physics defines.

* style

>.<

(sigh) null checks can be simplified style/standards update.

* update

switching from TT as a class to TT as a MonoBehaviour.

* fix

This fixes an issue where a server instance parenting with AllowOwnerToParent set would send to itself.
Adding check within the NetworkObject parenting logic to pass through to the server portion of sending the parenting message.

* test

adding some validation tests for the updates made in #3013

* update

adding changelog entries

* fix

This fixes the two failing tests within NetworkTransformAnticipationTest and updates AnticipatedNetworkTransform to account for #3013's updates/changes.

* fix

Removing the added base.OnInspectorGUI() call (was causing duplicated properties)

* update

Assure the callouts and call-to-action buttons show up outside of the Foldout Group when extending NetworkManager.
Assure that the NetworkManager has no Foldout Group when it is not extended.