fix: scene loading event not synchronizing in scene objects (#3096)
* fix

This fixes the issue with in-scene placed NetworkObjects not being synchronized properly when the session owner generates a scene event type load. The service is not considered an observer and when sending a scene event type load it should send all spawned objects.

* update

adding changelog entry

* update

adding associated PR to changelog entry

* update

covering synchronization too.