fix: networkvariablebase not being reinitialized if networkobject persists between sessions (#3181)
* fix

This will "re-initialize" NetworkVariableBase derived classes when the instance is not destroyed and repurposed.

* update

adding changelog entry

* update

Adding PR number to changelog entry

* fix

? This test needs more comments.

* update

Changing the approach to this fix while also cleaning up the initialize process so it can be invoked when the derived class is not yet fully spawned (or just recently instantiated) without potentially assigning the wrong NetworkManager instance. It is being invoked multiple times anyway, so this just gradually initializes all of the required elements until it is finally considered "initialized".
This also prevents OnInitialize from being invoked multiple times during this process and OnInitialize is only invoked as a last step to being considered "initialized".

Added a Deinitialize step to NetworkVariableBase when despawning a NetworkObject where the initialized status is reset. This allows for the next spawning of the NetworkObject (if in-scene placed or pooled) to run through the initialization process.

* test

reverting the changes to WhenServerChangesSmoothValue_ValuesAreLerped as the recent update resolves the issue from the previous fix.

* update

Removing exception.
Adjusting when we update the last time.
Logging warning if log level is developer and there is no NetworkTimeSystem.

* update

remove the try catch because evidently it is the only way we can catch certain exceptions during unit testing when duplicating the value (i.e. serializing).

* test

This validates the fix for resetting the NetworkVariableBase.LastUpdated property when a NetworkObject is persisted (i.e. recycled via in-scene placed NetworkObject or object pools).

Also minor update to prevent the integration test from logging a blank entry if there has been no logs added.

* test

The first iteration should ignore the time delta as that could be impacted by the test or VM running the test.