fix: selecting everything selects session owner only (#3305)
* fix

This fixes the issue with setting flags on NetworkObjects and the default Everything/All flag causing session owner to be grouped with the rest of the flags. Now all will initially select all flags but the session owner but if the all flag is selected again it toggles to just the SessionOwner flag.

This also assures that all root in-scene placed NetworkObjects have the Distributable or SessionOwner flag set (if neither then it defaults to Distributable).

* fix

Fixing some logic that would only skip over children if it was logging a message.

* update

adding changelog entries.

* style

adding API comments

* style

Adding one for "None".

* update

Cleaning up some of the script.

* style

adding comments