ci: [2.X] CI failures fixes (#3302)
This PR addresses some test failures in new CI definition (present on
develop-2.0.0 branch)

1. **Code Coverage test failure** --> This test was failing because
packed package couldn't be find. This steams from the fact that the path
was wrongly pointing to different place because package pack job was
modified to fix this.
2. **minimalproject standards failure** --> We were focusing before on
making sure that testproject is conforming to standards (which also
means that NGO package conforms to standards) but minimalproject has few
errors which we aim to fix in this PR
3. **testproject-tools-integration standards failure** --> We were
focusing before on making sure that testproject is conforming to
standards (which also means that NGO package conforms to standards) but
testproject-tools-integration has few errors which we aim to fix in this
PR

---------

Co-authored-by: NoelStephensUnity <noel.stephens@unity3d.com>