update
Exposing internal access to position and rotation interpolators for future tests and tools.
Noticed the slerp for full precision rotation was only being applied when using Lerp, now it is for all interpolation types.
Making the time to approximately reach adjustment be the default when there is no pending item to be consumed and slightly sooner when there is.
Removing the adjustment to the slerp smoothing as this should be a consistantly applied value for all interpolation types.