fix
This update resolves the issue where the legacy lerp result was not yielding the same result as the original lerp.
This update also resolves a secondary issue discovered by the services group where the final interpolation (i.e. last measurement in the queue) was not using a high enough precision value when determining if it reached its final destination point.
This update also includes some inlining additions along with a spelling mistake with the IsApproximately method.
It also updates the XML API regarding Legacy lerp to let users know that it does not use the tick latency value when calculating the ticksAgo value.