fix: LegacyLerp was using ticklatency as opposed to legacy ticksAgo (#3381)
This update resolves the issue where the legacy lerp result was not
yielding the same result as the original lerp. This update also resolves
a secondary issue discovered by the services group where the final
interpolation (i.e. last measurement in the queue) was not using a high
enough precision value when determining if it reached its final
destination point. This update also includes some inlining additions
along with a spelling mistake with the IsApproximately method. It also
updates the XML API regarding Legacy lerp to let users know that it does
not use the tick latency value when calculating the ticksAgo value.


## Changelog
NA

## Testing and Documentation

- No tests have been added.
- Includes edits to existing API XML documentation.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->