fix: Don't accept invalid connections in UnityTransport.Send (#3382)
This avoids the error message reported in
[MTTB-1143](https://jira.unity3d.com/browse/MTTB-1143).

The `UnityTransport.Send` method would happily accept invalid or stale
connections, which would lead to the allocation of a `BatchedSendQueue`
structure and to an error message when the batched sends are passed on
to the driver.

In MTTB-1143 we'd get a send on client ID 0, but that's not a valid
client ID in `UnityTransport`. My guess is that there's something
somewhere (possibly in Boss Room) triggering a send to the server after
the connection has closed (which reverts `UnityTransport.ServerClientId`
to its default value of 0). Of course, ideally there'd never be such a
send-after-disconnect and that's what should be ultimately addressed.
But the bug is quite hard to reproduce, so as a stopgap until we have
something better I'm making a fix that only avoids the last and more
user-visible consequences of the issue.