refactor: Minor cleanups in Unity Transport (#3377)
Just a few minor cleanups in Unity Transport:

- Remove the internal `NetworkDriver` accessor. There's a public
`GetNetworkDriver` nowadays that can be used for the same purpose.
- Add a `using` directive for transport errors which avoids using the
full (and long) type name.
- Remove the internal "state" member. The same information can be
gleaned by looking at the driver's state and the value of
`m_ServerClientId`. Plus, the `Listening` state did not have the same
meaning as it does in NGO itself, which could lead to confusion.
- Simplify `ErrorUtilities` to avoid the useless constants and only
providing strings for errors users can actually do something about. Also
moved it to the bottom of the file. The top is too prime real estate for
such a minor utility. :P