fix: In-Scene Placed Object Parenting, Serialization Order, and Transform Parent Detection (#3387)
This resolves issues with in-scene placed NetworkObject parenting where
the original parent has changed. This resolves the issue where loading a
scene would not order the serialization of loaded in-scene placed
NetworkObjects based on their parent-child hierarchy.

This also resolves an issue where a `NetworkTransform` would think it
was parented if nested under anything as well as removes the spamming
error message when parenting a `NetworkObject` with `NetworkTransform`
and using local space for synchronization (can be handled in
documentation as to why this is a bad idea but there may be edge cases
where a user might want to do this anyway).


[MTT-11883](https://jira.unity3d.com/browse/MTT-11883)

<!-- Add RFC link here if applicable. -->

## Changelog

- Fixed: Issue where in-scene placed `NetworkObjects` could fail to
synchronize its transform properly (especially without a
`NetworkTransform`) if their parenting changes from the default when the
scene is loaded and if the same scene remains loaded between network
sessions while the parenting is completely different from the original
hierarchy.
- Changed: The scene loading event serialization order for in-scene
placed `NetworkObject`s to be based on their parent-child hierarchy.
- Changed: Removing the error message when a `NetworkObject` with
`NetworkTransform` is parented and placed in local space when using
rigidbody for motion.


## Testing and Documentation

- Includes no additional tests (_requires manual testing due to scene
loading constraints_).
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->